atulya-integrations/pydantic-ai/pyproject.toml
[project]
name = "atulya-pydantic-ai"
version = "0.8.3"
description = "Pydantic AI integration for Atulya - persistent memory tools for AI agents"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Anurag Atulya", email = "support@eightengine.com" }
]
keywords = [
    "ai",
    "memory",
    "pydantic-ai",
    "agents",
    "atulya",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic-ai-slim>=1.0.0",
    "atulya-client>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/eight-atulya/atulya"
Documentation = "https://github.com/eight-atulya/atulya/tree/main/atulya-integrations/pydantic-ai"
Repository = "https://github.com/eight-atulya/atulya"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["atulya_pydantic_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]
